+++ hat/assets/js/apps/Iaso/constants/urls.ts
@@ -487,16 +487,7 @@ export const baseRouteConfigs: Record<string, RouteConfig> = {
     },
     assignments: {
         url: 'planning/assignments',
-        params: [
-            'accountId',
-            'planningId',
-            'team',
-            'baseOrgunitType',
-            'parentOrgunitType',
-            'tab',
-            'order',
-            'search',
-        ],
+        params: ['accountId', 'planningId'],
     },
     teams: {
         url: 'settings/users/teams',
