+++ hat/assets/js/apps/Iaso/domains/assignments/types/assigment.ts
@@ -3,11 +3,6 @@ import { Shape } from '../../orgUnits/types/shapes';
 export type AssignmentParams = {
     planningId: string;
     team?: string;
-    baseOrgunitType?: string;
-    parentOrgunitType?: string;
-    tab?: string;
-    order?: string;
-    search?: string;
 };
 
 type OrgUnitDetails = {
